+++ src/pages/ManageUser/ManageUser.css
@@ -8,11 +8,6 @@
 .manage-user-container .tabs {
     margin: 10px 55px;
 }
-.manage-user-container .tabs span {
-    margin-right: 20px;
-    cursor: pointer;
-    font-weight: 400;
-}
 .manage-user-container .users-content .btn-add-new button{
     border: none;
     margin: 15px 0px;
